backend/requirements/base.txt
Django>=6.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.3
django-filter>=25.1
django-environ>=0.11
psycopg2-binary>=2.9
gunicorn>=22.0
